Create prototype repository
Create index.html. Its an html file which contains a scroll-spy with
bootstrap, so the page can be scrolled while the page monitors the
scrolling and knows in which section it is. In section1, we use jquery
and d3 to call a graph once the document is ready.
Create section4.js. This is where the d3 draws a scatter graph using
data from a local dataset. There are two axes drawn normally and they
work correctly. The point [0,0] was added and then not displayed so the
axes would show the origin.
Create main.css, which is the styles file.